requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
oauthlib==2.0.7
pika==0.11.2
pyaml==17.12.1
PySocks==1.6.8
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tweepy==3.6.0
tweet-preprocessor==0.5.0
urllib3==1.22
flask
flask-cors
pymongo
numpy
gensim
keras
tensorflow
